Refactor build to extract common test utils
