+++ project/ScalafixBuild.scala
@@ -174,17 +174,18 @@ object ScalafixBuild extends AutoPlugin with GhpagesKeys {
       "clean" ::
         "scalafix/publishSigned" ::
         "scalafix211/publishSigned" ::
+        "scalafix-sbt013/publishSigned" ::
         "scalafix-sbt/publishSigned" ::
-        s"^^ $sbt1 scalafix-sbt/publishSigned" ::
+        "sonatypeReleaseAll" ::
         s
     },
     commands += Command.command("ci-sbt") { s =>
       // scripted tests don't work in sbt 1.0 yet because we run Sbt1
-      s"^^ $sbt1 scalafix-sbt/publishLocal" ::
+      "scalafix-sbt/publishLocal" ::
         s
     },
     commands += Command.command("ci-sbt-sbt013") { s =>
-      "scalafix-sbt/scripted" ::
+      "scalafix-sbt013/scripted" ::
         s
     },
     commands += Command.command("ci-fast-212") { s =>
